\name{names.Sardinia}
\alias{names.Sardinia}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
names.Sardinia
}
\description{
data frame/data.table; it contains Code, integer representing the id of the community (elementary area or municipality) and com.name, character, containing the community name. The Code must be positive. In order to differ the community name by the lma with the same name, the community is expressed as  first letter in uppercase and the remaining letters in lowercase, whereas the lma name is all in uppercase.  
}
\usage{
data(names.Sardinia)
}
%- maybe also 'usage' for other objects documented here.
\author{
Daniela Ichim, Luisa Franconi, Michele D'Alo'
}
