\name{Sardinia}
\alias{Sardinia}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sardinia
}
\description{data frame/data.table containing the commuting flows between municipalities in Sardinia Island (Italy) from 2001 Italian Population Census (the complete data set http://www.istat.it/it/archivio/157423).  Each row corresponds to an observation i.e. a flow and each column corresponds to a variable.  The variables are:  community_live, integer, id number of the community where the employees live (elementary zone of residence), community_work, integer, containing the id number of the community where the employees work (elementary zone of work) and amount, numeric containing the number of employees commuting between community_live and community_work (direction is important). 
}
\usage{
data(Sardinia)
}
\author{
Daniela Ichim, Luisa Franconi and Michele D'Alo' 
}
