\name{shpSardinia}
\alias{shpSardinia}
\docType{data}
\title{shpSardinia

}
\description{
object of class sf containing that holds polygons with attributes for each community in Sardinia island as from Italian Population Census 2001.    
}
\usage{data("shpSardinia")}
\source{
The data is taken from the shape files (in generalised form) of the whole of Italian municipalities at the 2001 Populalation Census published by Istat: http://www.istat.it/it/archivio/124086 
}
\examples{
\dontrun{
data(shpSardinia)
plot(shpSardinia)
}
## maybe str(shpSardinia) ; ...
}
\keyword{datasets}
